Refatorando checkpoint

+++ checkpointFinal/src/main/java/com/example/checkpointFinal/repositories/ProductRepository.java
@@ -3,5 +3,6 @@ package com.example.checkpointFinal.repositories;
 import com.example.checkpointFinal.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product,Integer> {
+public interface ProductRepository extends JpaRepository<Product,Long> {
 }
+
